face-service/requirements.txt
face_recognition
opencv-python
numpy
flask
flask-cors
pillow
firebase-admin
